requirements.txt
# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Machine Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# CLIP and Vision Models
open_clip_torch>=2.20.0

# Vector Search
faiss-cpu>=1.7.4

# Dataset and Data Processing
datasets>=2.14.0
huggingface-hub>=0.19.0

# Image Processing
Pillow>=10.0.0

# Numerical Computing
numpy>=1.24.0

# Optional but Recommended
# matplotlib>=3.7.0  # Only if you need visualization features
pydantic
pydantic-settings